--- src/common.t.ts
@@ -1,2 +1,0 @@
-import type * as t from './types';
-export type { t };
