+++ src/common/index.ts
@@ -1,7 +1,6 @@
+import type * as t from '../types';
 import * as Util from './util';
 
 export * from './const.hashes';
 export * from './libs';
-export { Util };
-
-export { type t } from '../common.t';
+export { Util, type t };
